Hopefully fixing commit b9510bcaa5da65288e53acfc1738bc48eae6562e as the GLPI installation doesn't work properly anymore.
Why doesn't it work: This is more of an assumption and might be something else. However, the installation worked properly with the example docker-compose.yml where no volumes were mapped, but didn't work, when the volume mapping /var/www/html/glpi/:/var/www/html/glpi was included. It could be that docker created the volume folder, the container was started, checked if the glpi directory was present (which it was) and did the additional check, if 'ls /var/www/html/glpi' did't just return 'plugins'. This check might failed since the glpi directory existed, but 'ls /var/www/html/glpi' returned an empty string and thus considered GLPI to be installed.